file will not be downloaded again if it already exists

+++ src/helpers/blog_helper.rb
@@ -66,13 +66,15 @@ module BlogHelper
     dir = Pathname("site/images/") + Date.today.strftime("%Y%m%d")
     FileUtils.mkdir_p dir
 
-    if uri =~ /^[a-z]{3,5}:\/{2}.*\..{2,5}\//
-      # it's an URL, so we need to download the picture first
-      return nil unless system "wget #{uri} -O #{dir+name}"
-    else
-      # it's a local file, so we need to copy it into the blog images directory
-      return nil unless File.exist? uri
-      FileUtils.cp(uri, dir+name)
+    unless File.exist? dir+name
+      if uri =~ /^[a-z]{3,5}:\/{2}.*\..{2,5}\//
+        # it's an URL, so we need to download the picture first
+        return nil unless system "wget #{uri} -O #{dir+name}"
+      else
+        # it's a local file, so we need to copy it into the blog images directory
+        return nil unless File.exist? uri
+        FileUtils.cp(uri, dir+name)
+      end
     end
 
     return nil unless system "src/helpers/create_thumbs.sh #{dir+name}"
